refactor: componente GoalChartCard e separa funlçoes para reaproveitamento
