+++ src/components/Features/Comercial/GoalChartCard.tsx
@@ -1,18 +1,17 @@
 import { GoalChart } from '@/components/Charts/GoalChart';
 import GoalChartSkeleton from '@/components/Skeletons/GoalChartSkeleton';
+import { generateColor } from '@/functions/charts/generateColor';
+import { GroupedResults, groupResultsByUsuario } from '@/functions/charts/groupResultsByUsuario';
 
-type Result = {
+export type Result = {
     id: string;
     credor: string;
     usuario: string;
     meta_1: number;
     meta_2: number;
     meta_3: number;
+    valor_liquido_disponivel: number;
     precatorio_fechado_em: string;
-};
-
-type GroupedResults = {
-    [usuario: string]: Result[];
 };
 
 interface GoalChartCardProps {
@@ -22,36 +21,6 @@ interface GoalChartCardProps {
 export function GoalChartCard({ results }: GoalChartCardProps) {
     if (!results || results.length === 0) {
         return <GoalChartSkeleton />;
-    }
-
-    function groupByUsuario(results: Result[]) {
-        return results.reduce((acc, result) => {
-            const { usuario } = result;
-            if (!acc[usuario]) {
-                acc[usuario] = [];
-            }
-            acc[usuario].push(result);
-            return acc;
-        }, {} as GroupedResults);
-    }
-
-    function generateColor(color: 'red' | 'yellow' | 'green', index: number, steps: number) {
-        const baseHues = {
-            red: 0,
-            yellow: 60,
-            green: 120,
-        };
-
-        const baseHue = baseHues[color];
-        const saturation = 70;
-
-        const minLightness = 25;
-        const maxLightness = 55;
-        const stepSize = (maxLightness - minLightness) / (steps - 1);
-
-        const lightness = maxLightness - index * stepSize;
-
-        return `hsl(${baseHue}, ${saturation}%, ${lightness}%)`;
     }
 
     function groupChartDataByGoal(
@@ -72,7 +41,7 @@ export function GoalChartCard({ results }: GoalChartCardProps) {
         }));
     }
 
-    const groupedResults = groupByUsuario(results);
+    const groupedResults = groupResultsByUsuario(results);
 
     const meta1Data = groupChartDataByGoal(groupedResults, 'meta_1', 'red');
     const meta2Data = groupChartDataByGoal(groupedResults, 'meta_2', 'yellow');
